completed GuardAI
fixed scripts so that Guard walks on a waypoint patrol path, fixed so that if guard sees player either by a distance or from his surrounding that he chases the enemy, also when he hits the enemy he gives out a message so that when he collides, something can happen. made XR Orgin as a prefab as well as the guard. there is also an interface script that reads on the player and sends a message to the guard that the ray hit him. added a tag enemy on the guard.